pyproject.toml
[tool.poetry]
name = "qualification"
version = "0.1.0"
description = ""
authors = ["Vlad Morozov <morozov90vlad@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
django-dump-load-utf8 = "^0.0.4"
django-cors-headers = "^4.4.0"
drf-yasg = "^1.21.7"
setuptools = "^71.0.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
mypy = "^1.11.0"
coverage = "^7.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'
warn_unreachable = true
ignore_missing_imports = true

[tool.black]
line-length = 119

[tool.isort]
line_length = 119